Separate styles from components, add default assigning, use object destructuring and etc.
