+++ src/App.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ButtonMenu from './components/ButtonMenu';
+import ButtonMenu from './components/ButtonMenu/ButtonMenu';
 import logoName from './images/logo-name.svg';
 import logoCircle from './images/circle-logo.svg';
