+++ src/api/index.js
@@ -8,3 +8,15 @@ export const getAllBeersForCategory = (category) =>
 
 export const getAllCategories = () =>
   Array.from(new Set(beerList.map((beer) => beer.category))).sort();
+
+export const getFullInfoKeys = (beerId) => {
+  const beerInformation = getAllBeerInformation(beerId);
+  const shortInfo = ['id', 'name', 'image', 'description', 'price', 'amount'];
+  const fullInfo = [];
+  Object.keys(beerInformation).forEach((keyName) => {
+    if (!shortInfo.includes(keyName)) {
+      fullInfo.push(keyName);
+    }
+  });
+  return fullInfo;
+};
